adding caption to meme works now

+++ app/src/main/java/com/bolnizar/memro/events/OpenMemeCaption.java
@@ -4,9 +4,9 @@ package com.bolnizar.memro.events;
  * Created by BoldijarPaul on 18/08/16.
  */
 public class OpenMemeCaption {
-    public OpenMemeCaption(int memeId) {
+    public OpenMemeCaption(Long memeId) {
         this.memeId = memeId;
     }
 
-    public final int memeId;
+    public final Long memeId;
 }
